chore: some more refactoring
