+++ src/components/EarnFilterBar/EarnFilterBar.stories.tsx
@@ -1,6 +1,9 @@
 import type { Meta, StoryObj } from '@storybook/nextjs-vite';
 import { useState } from 'react';
-import { EarnFilteringContext } from '../../app/ui/earn/EarnFilteringContext';
+import {
+  EarnFilteringContext,
+  SortByOptions,
+} from '../../app/ui/earn/EarnFilteringContext';
 import { EarnCardVariant } from '../Cards/EarnCard/EarnCard.types';
 import { EarnFilterBar } from './EarnFilterBar';
 
@@ -30,8 +33,11 @@ type Story = StoryObj<typeof meta>;
 
 const mockContextValue = () => {
   const [showForYou, setShowForYou] = useState(false);
+  const [sortBy, setSortBy] = useState<SortByOptions>(SortByOptions.APY);
 
   return {
+    sortBy,
+    setSortBy,
     filter: {},
     updateFilter: () => {},
     showForYou,
