+++ src/components/EarnFilterBar/components/EarnFilterSort.tsx
@@ -1,21 +1,14 @@
 import { useTranslation } from 'react-i18next';
-import { SingleSelect } from '../core/SingleSelect/SingleSelect';
-import { Select } from '../core/form/Select/Select';
-import { SelectVariant } from '../core/form/Select/Select.types';
+import { Select } from '../../core/form/Select/Select';
+import { SelectVariant } from '../../core/form/Select/Select.types';
+import {
+  SortByOptions,
+  useEarnFiltering,
+} from 'src/app/ui/earn/EarnFilteringContext';
 
-// TODO: migrate to backend's typing
-export enum SortByOptions {
-  APY = 'apy',
-  TVL = 'tvl',
-}
-
-type Props = {
-  sortBy: SortByOptions;
-  setSortBy: (sortBy: SortByOptions) => void;
-};
-
-export const EarnFilterSort: React.FC<Props> = ({ sortBy, setSortBy }) => {
+export const EarnFilterSort: React.FC = () => {
   const { t } = useTranslation();
+  const { sortBy, setSortBy } = useEarnFiltering();
 
   const handleChange = (value: string) => {
     setSortBy(value as SortByOptions);
